\name{DVCP-package}
\alias{DVCP-package}
\alias{DVCP}
\docType{package}
\title{
\packageTitle{DVCP}
}
\description{
\packageDescription{DVCP}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{DVCP}
\packageIndices{DVCP}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{DVCP}

Maintainer: \packageMaintainer{DVCP}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
